requirements.txt
click==8.1.3
Flask==3
itsdangerous==2.1.2
Jinja2==3.1.2
Werkzeug==3
tensorflow==2.19.0
pillow==10.4.0
